TH-211 Refactor: Extract duplicated code to mixin class
